Updated source without useless vars.
Free Software Free Society

To support commits by ctubio,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/lib/db.h
@@ -8,7 +8,7 @@ namespace K {
     public:
       static void main(Local<Object> exports) {
         Isolate* isolate = exports->GetIsolate();
-        dbFpath = string("/data/db/K.").append(to_string((int)CF::cfExchange())).append(".").append(to_string((int)CF::cfBase())).append(".").append(to_string((int)CF::cfQuote())).append(".db");
+        dbFpath = string("/data/db/K.").append(to_string((int)CF::cfExchange())).append(".").append(to_string(CF::cfBase())).append(".").append(to_string(CF::cfQuote())).append(".db");
         if (sqlite3_open(dbFpath.data(), &db)) { cout << FN::uiT() << sqlite3_errmsg(db) << endl; exit(1); }
         cout << FN::uiT() << "DB " << dbFpath << " loaded OK." << endl;
         NODE_SET_METHOD(exports, "dbLoad", DB::_load);
